Move algorithm and parsing to lib directory
