+++ lib/camper.py
@@ -13,16 +13,12 @@ class Camper:
         except ValueError:
             return len(self.preferences)
 
-
-
 def merge_objects(preferences, activities, histories):    # These are all list objects
     """Takes in lists of histories, activities, preferences, and gives back a list of campers with all traits in one place"""
     histories_dict = {history.name: (history.past_activities, history.past_preferences) for history in histories}
     activities_dict = {activity.name: activity for activity in activities}
     campers = [create_camper(preference, histories_dict, activities_dict) for preference in preferences]
     return campers
-
-
 
 def create_camper(preference, histories_dict, activities_dict):
     """Takes the preference object to be made a camper and dictionaries of histories and activities, outputs a Camper object"""
