+++ lib/xls/parsing/preference.py
@@ -1,4 +1,4 @@
-from xls_parsing.common import locate_header
+from .common import locate_header
 
 class Preference:
     def __init__(self, name, edah, bunk, preferences):
