+++ lib/xls/validation.py
